+++ tests/test_spying.py
@@ -12,7 +12,7 @@ from .common import DerivedClass, Proxy, SomeClass
 from .utils import assert_raises
 
 
-class TestSpying:
+class SpyingTestCase:
     def test_spy_should_cache_asserts(self) -> None:
         assert1 = mocker(SomeClass).spy("instance_method")
         assert2 = mocker(SomeClass).spy("instance_method")